// data/github/current/ui/packages/issues-bulk-actions/__tests__/__generated__/AddToProjectsBulkActionTestQuery.graphql.ts
export type AddToProjectsBulkActionTestQuery$variables = {
  ids: ReadonlyArray<string>;
};
export type AddToProjectsBulkActionTestQuery$data = {
  readonly nodes: ReadonlyArray<{
    readonly id?: string;
    readonly projectItemsNext?: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly id: string;
          readonly project: {
            readonly " $fragmentSpreads": FragmentRefs<"ProjectPickerProject">;
          };
        } | null | undefined;
      } | null | undefined> | null | undefined;
    } | null | undefined;
  } | null | undefined>;
};
export type AddToProjectsBulkActionTestQuery = {
  response: AddToProjectsBulkActionTestQuery$data;
  variables: AddToProjectsBulkActionTestQuery$variables;
};